% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dollars_raised.R
\name{dollars_raised}
\alias{dollars_raised}
\title{dollars_raised}
\usage{
dollars_raised(unit, from, to)
}
\arguments{
\item{unit}{Unit code}

\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
The function outputs the total dollars raised and number of gifts secured by a unit
in a given timeframe. The data is pulled from proposals with an assignment to the
identified unit.
}
\examples{
dollars_raised(unit = 'EN', from = 20140101, to = 20180630)
}
